Make it so it outputs best value for each subproblem
At the moment it seems like it completely solves a given problem and outputs the best value for each subproblem.
The only current issue is I don't know why it finds the subproblem cutting a string of length 4 at position 2, since the problem I give it doesn't feel like that would ever happen.
Next step is to add the order in which each cut happens.